move out code handling logic to another component
